+++ src/controller.ts
@@ -1,6 +1,7 @@
 import type { Disposable } from '@type/disposable';
-import type { Maidr } from '@type/maidr';
+import type { Maidr } from '@type/grammar';
 import { Context } from '@model/context';
+import { Figure } from '@model/plot';
 import { AudioService } from '@service/audio';
 import { AutoplayService } from '@service/autoplay';
 import { BrailleService } from '@service/braille';
@@ -12,6 +13,7 @@ import { KeybindingService } from '@service/keybinding';
 import { NotificationService } from '@service/notification';
 import { ReviewService } from '@service/review';
 import { SettingsService } from '@service/settings';
+import { LocalStorageService } from '@service/storage';
 import { TextService } from '@service/text';
 import { store } from '@state/store';
 import { BrailleViewModel } from '@state/viewModel/brailleViewModel';
@@ -22,7 +24,6 @@ import { ViewModelRegistry } from '@state/viewModel/registry';
 import { ReviewViewModel } from '@state/viewModel/reviewViewModel';
 import { SettingsViewModel } from '@state/viewModel/settingsViewModel';
 import { TextViewModel } from '@state/viewModel/textViewModel';
-import { Figure } from '@type/plot';
 
 export class Controller implements Disposable {
   private readonly figure: Figure;
@@ -58,7 +59,7 @@ export class Controller implements Disposable {
 
     this.displayService = new DisplayService(this.context, plot, reactContainer);
     this.notificationService = new NotificationService();
-    this.settingsService = new SettingsService(this.displayService);
+    this.settingsService = new SettingsService(new LocalStorageService(), this.displayService);
 
     this.audioService = new AudioService(this.notificationService, this.context.state);
     this.brailleService = new BrailleService(this.context, this.notificationService, this.displayService);
@@ -143,6 +144,8 @@ export class Controller implements Disposable {
     this.figure.subplots.forEach(subplotRow => subplotRow.forEach((subplot) => {
       subplot.addObserver(this.textService);
       subplot.addObserver(this.audioService);
+      subplot.addObserver(this.brailleService);
+      subplot.addObserver(this.highlightService);
       subplot.traces.forEach(traceRow => traceRow.forEach((trace) => {
         trace.addObserver(this.audioService);
         trace.addObserver(this.brailleService);