+++ src/model/line.ts
@@ -1,18 +1,15 @@
-import type { MaidrLayer } from '@type/maidr';
-import type { AudioState, TextState } from '@type/state';
-import type { LinePoint } from './grammar';
+import type { LinePoint, MaidrLayer } from '@type/grammar';
+import type { AudioState, BrailleState, TextState } from '@type/state';
 import { Constant } from '@util/constant';
 import { Svg } from '@util/svg';
-import { AbstractTrace } from './plot';
+import { AbstractTrace } from './abstract';
 
 const TYPE = 'Type';
 const SVG_PATH_LINE_POINT_REGEX = /[ML]\s*(-?\d+(\.\d+)?)\s+(-?\d+(\.\d+)?)/g;
 
-export class LinePlot extends AbstractTrace<number> {
+export class LineTrace extends AbstractTrace<number> {
   private readonly points: LinePoint[][];
   private readonly lineValues: number[][];
-
-  protected readonly brailleValues: string[][];
   protected readonly highlightValues: SVGElement[][] | null;
 
   private readonly min: number[];
@@ -27,16 +24,11 @@ export class LinePlot extends AbstractTrace<number> {
     this.min = this.lineValues.map(row => Math.min(...row));
     this.max = this.lineValues.map(row => Math.max(...row));
 
-    this.brailleValues = this.mapToBraille(this.lineValues);
     this.highlightValues = this.mapToSvgElements(layer.selectors as string[]);
   }
 
   public dispose(): void {
     this.points.length = 0;
-    this.lineValues.length = 0;
-
-    this.brailleValues.length = 0;
-    this.highlightValues && (this.highlightValues.length = 0);
 
     this.min.length = 0;
     this.max.length = 0;
@@ -58,6 +50,18 @@ export class LinePlot extends AbstractTrace<number> {
     };
   }
 
+  protected braille(): BrailleState {
+    return {
+      empty: false,
+      id: this.id,
+      values: this.lineValues,
+      min: this.min,
+      max: this.max,
+      row: this.row,
+      col: this.col,
+    };
+  }
+
   protected text(): TextState {
     const point = this.points[this.row][this.col];
     const fillData = point.fill
@@ -71,74 +75,6 @@ export class LinePlot extends AbstractTrace<number> {
     };
   }
 
-  private mapToBraille(data: number[][]): string[][] {
-    const braille = new Array<Array<string>>();
-
-    for (let row = 0; row < data.length; row++) {
-      braille.push(new Array<string>());
-
-      const range = (this.max[row] - this.min[row]) / 4;
-      const low = this.min[row] + range;
-      const medium = low + range;
-      const mediumHigh = medium + range;
-      const high = medium + range;
-
-      for (let col = 0; col < data[row].length; col++) {
-        if (data[row][col] <= low && col - 1 >= 0 && data[row][col - 1] > low) {
-          if (data[row][col - 1] <= medium) {
-            braille[row].push('⢄');
-          } else if (data[row][col - 1] <= mediumHigh) {
-            braille[row].push('⢆');
-          } else if (data[row][col - 1] > mediumHigh) {
-            braille[row].push('⢇');
-          }
-        } else if (data[row][col] <= low) {
-          braille[row].push('⣀');
-        } else if (col - 1 >= 0 && data[row][col - 1] <= low) {
-          if (data[row][col] <= medium) {
-            braille[row].push('⡠');
-          } else if (data[row][col] <= mediumHigh) {
-            braille[row].push('⡰');
-          } else if (data[row][col] > mediumHigh) {
-            braille[row].push('⡸');
-          }
-        } else if (
-          data[row][col] <= medium
-          && col - 1 >= 0
-          && data[row][col - 1] > medium
-        ) {
-          if (data[row][col - 1] <= mediumHigh) {
-            braille[row].push('⠢');
-          } else if (data[row][col - 1] > mediumHigh) {
-            braille[row].push('⠣');
-          }
-        } else if (data[row][col] <= medium) {
-          braille[row].push('⠤');
-        } else if (col - 1 >= 0 && data[row][col - 1] <= medium) {
-          if (data[row][col] <= mediumHigh) {
-            braille[row].push('⠔');
-          } else if (data[row][col] > mediumHigh) {
-            braille[row].push('⠜');
-          }
-        } else if (
-          data[row][col] <= mediumHigh
-          && col - 1 >= 0
-          && data[row][col - 1] > mediumHigh
-        ) {
-          braille[row].push('⠑');
-        } else if (data[row][col] <= mediumHigh) {
-          braille[row].push('⠒');
-        } else if (col - 1 >= 0 && data[row][col - 1] <= mediumHigh) {
-          braille[row].push('⠊');
-        } else if (data[row][col] <= high) {
-          braille[row].push('⠉');
-        }
-      }
-    }
-
-    return braille;
-  }
-
   private mapToSvgElements(selectors?: string[]): SVGElement[][] | null {
     if (!selectors || selectors.length !== this.lineValues.length) {
       return null;
@@ -146,13 +82,12 @@ export class LinePlot extends AbstractTrace<number> {
 
     const svgElements = new Array<Array<SVGElement>>();
     for (let r = 0; r < selectors.length; r++) {
-      const domElements = Array.from(document.querySelectorAll<SVGElement>(selectors[r]));
-      if (domElements.length !== 1) {
+      const lineElement = Svg.selectElement(selectors[r], false);
+      if (!lineElement) {
         return null;
       }
 
       const coordinates = new Array<LinePoint>();
-      const lineElement = domElements[0];
       if (lineElement instanceof SVGPathElement) {
         const pathD = lineElement.getAttribute(Constant.D) || Constant.EMPTY;
         let match = SVG_PATH_LINE_POINT_REGEX.exec(pathD);
@@ -172,13 +107,12 @@ export class LinePlot extends AbstractTrace<number> {
         return null;
       }
 
-      const style = window.getComputedStyle(lineElement);
       const linePointElements = new Array<SVGElement>();
       for (const coordinate of coordinates) {
         if (Number.isNaN(coordinate.x) || Number.isNaN(coordinate.y)) {
           return null;
         }
-        linePointElements.push(Svg.createCircleElement(coordinate.x, coordinate.y, style, lineElement));
+        linePointElements.push(Svg.createCircleElement(coordinate.x, coordinate.y, lineElement));
       }
       svgElements.push(linePointElements);
     }