+++ src/ui/pages/Braille.tsx
@@ -1,25 +1,18 @@
-import type { FormEvent } from 'react';
 import { useViewModel, useViewModelState } from '@state/hook/useViewModel';
+import { DomEventType } from '@type/event';
 import { Constant } from '@util/constant';
-import React, { useEffect, useRef } from 'react';
+import React, { useEffect, useId, useRef } from 'react';
 
 const Braille: React.FC = () => {
+  const id = useId();
   const viewModel = useViewModel('braille');
   const { value, index } = useViewModelState('braille');
 
   const brailleRef = useRef<HTMLTextAreaElement>(null);
   const lastIndexRef = useRef<number>(index);
 
-  useEffect(() => {
-    if (brailleRef.current) {
-      brailleRef.current.selectionStart = index;
-      brailleRef.current.selectionEnd = index;
-      lastIndexRef.current = index;
-    }
-  }, [index]);
-
-  const handleSelectionChange = (event: FormEvent<HTMLTextAreaElement>): void => {
-    const textArea = event.currentTarget;
+  const handleSelectionChange = (event: Event): void => {
+    const textArea = event.target as HTMLTextAreaElement;
     const newIndex = textArea.selectionStart;
 
     if (newIndex !== lastIndexRef.current) {
@@ -28,13 +21,35 @@ const Braille: React.FC = () => {
     }
   };
 
+  useEffect(() => {
+    const textArea = brailleRef.current;
+    if (!textArea) {
+      return;
+    }
+
+    textArea.addEventListener(DomEventType.SELECTION_CHANGE, handleSelectionChange);
+    return () => {
+      textArea.removeEventListener(DomEventType.SELECTION_CHANGE, handleSelectionChange);
+    };
+  }, []);
+  useEffect(() => {
+    const textArea = brailleRef.current;
+    if (!textArea) {
+      return;
+    }
+
+    textArea.value = value;
+    textArea.selectionStart = index;
+    textArea.selectionEnd = index;
+    lastIndexRef.current = index;
+  }, [value, index]);
+
   return (
-    <div>
+    <div id={id}>
       <textarea
-        id={Constant.BRAILLE_TEXT_AREA}
+        id={`${Constant.BRAILLE_TEXT_AREA}-${id}`}
         ref={brailleRef}
         defaultValue={value}
-        onSelect={handleSelectionChange}
         autoCapitalize="off"
         autoFocus
       />