+++ src/ui/pages/Review.tsx
@@ -1,15 +1,17 @@
 import { useViewModelState } from '@state/hook/useViewModel';
 import { Constant } from '@util/constant';
-import React from 'react';
+import React, { useId } from 'react';
 
 const Review: React.FC = () => {
+  const id = useId();
   const { value } = useViewModelState('review');
 
   return (
-    <div>
+    <div id={id}>
       <input
-        id={Constant.REVIEW_INPUT}
+        id={`${Constant.REVIEW_INPUT}-${id}`}
         defaultValue={value}
+        onFocus={event => event.target.setSelectionRange(0, 0)}
         size={50}
         type="text"
         autoComplete="off"